refactor: move split_lines_by_points to split.py, remove index hashing
