+++ src/geogenalg/core/geometry.py
@@ -6,8 +6,8 @@
 from collections.abc import Callable, Iterable
 from copy import deepcopy
 from enum import Enum
-from itertools import chain, pairwise
-from math import atan2, degrees, isclose, pi, sqrt
+from itertools import chain
+from math import atan2, degrees, pi, sqrt
 from statistics import mean
 from typing import Literal, NamedTuple
 from warnings import warn
@@ -45,7 +45,6 @@ from geogenalg.core.exceptions import (
     GeometryTypeError,
     InvalidGeometryError,
 )
-from geogenalg.identity import hash_index_from_geometry
 
 
 class LineExtendFrom(Enum):
@@ -1666,151 +1665,3 @@ def smooth_around_connection_point_of_two_lines(  # noqa: C901
         smoothed_lines.append(remove_repeated_points(modified_line))
 
     return tuple(smoothed_lines)
-
-
-def split_lines_by_points(
-    lines_gdf: GeoDataFrame,
-    points_gdf: GeoDataFrame,
-    max_distance: float = 1.0,
-) -> GeoDataFrame:
-    """Split line geometries at locations nearest to nearby points.
-
-    Points within max_distance from a line are projected onto the line and
-    used as split locations.
-
-    Args:
-    ----
-        lines_gdf: GeoDataFrame containing LineString geometries.
-        points_gdf: GeoDataFrame containing Point geometries.
-        max_distance: Maximum snapping distance from a point to the line.
-
-    Returns:
-    -------
-        GeoDataFrame containing split LineString geometries.
-
-    """
-    points_sindex = points_gdf.sindex
-
-    result_rows = []
-
-    for row in lines_gdf.itertuples():
-        line: LineString = row.geometry
-
-        # Find only nearby candidate points using spatial index
-        candidate_idx = list(
-            points_sindex.intersection(
-                line.buffer(max_distance).bounds,
-            )
-        )
-
-        if not candidate_idx:
-            new_row = lines_gdf.loc[row.Index].copy()
-            result_rows.append(new_row)
-            continue
-
-        candidate_points = points_gdf.iloc[candidate_idx]
-
-        split_distances: list[float] = []
-
-        for point in candidate_points.geometry:
-            if line.distance(point) >= max_distance:
-                continue
-
-            distance_along = line.project(point)
-            split_distances.append(distance_along)
-
-        if not split_distances:
-            new_row = lines_gdf.loc[row.Index].copy()
-            result_rows.append(new_row)
-            continue
-
-        split_distances = sorted(set(split_distances))
-
-        segments = _split_line_at_distances(
-            line,
-            split_distances,
-        )
-
-        for segment in segments:
-            new_row = lines_gdf.loc[row.Index].copy()
-            new_row.geometry = segment
-
-            result_rows.append(new_row)
-
-    result = GeoDataFrame(
-        result_rows,
-        crs=lines_gdf.crs,
-    )
-
-    return hash_index_from_geometry(
-        result,
-        hash_prefix="split_lines_by_points_",
-    )
-
-
-def _split_line_at_distances(
-    line: LineString,
-    distances: list[float],
-) -> list[LineString]:
-    """Split a LineString at distances measured along the line.
-
-    Returns
-    -------
-        List of LineString segments resulting from the split.
-
-    """
-    if not distances:
-        return [line]
-
-    coords = list(line.coords)
-
-    result_segments: list[LineString] = []
-
-    current_coords = [coords[0]]
-
-    distance_iter = iter(distances)
-    next_split = next(distance_iter, None)
-
-    accumulated = 0.0
-
-    for start_coord, end_coord in pairwise(coords):
-        segment = LineString([start_coord, end_coord])
-
-        segment_length = segment.length
-        segment_end_distance = accumulated + segment_length
-
-        while (
-            next_split is not None and accumulated < next_split < segment_end_distance
-        ):
-            relative_distance = next_split - accumulated
-
-            split_point = segment.interpolate(relative_distance)
-
-            split_coord = (split_point.x, split_point.y)
-
-            current_coords.append(split_coord)
-
-            result_segments.append(LineString(current_coords))
-
-            current_coords = [split_coord]
-
-            next_split = next(distance_iter, None)
-
-        if next_split is not None and isclose(next_split, segment_end_distance):
-            current_coords.append(end_coord)
-
-            result_segments.append(LineString(current_coords))
-
-            current_coords = [end_coord]
-
-            next_split = next(distance_iter, None)
-
-        else:
-            current_coords.append(end_coord)
-
-        accumulated = segment_end_distance
-
-    if len(current_coords) > 1:
-        result_segments.append(LineString(current_coords))
-
-    return result_segments
