+++ test/core/test_geometry.py
@@ -72,7 +72,6 @@ from geogenalg.core.geometry import (
     smooth_around_ring_closing_vertex,
     snap_to_closest_vertex_or_segment,
     split_linear_geometry,
-    split_lines_by_points,
 )
 
 
@@ -2607,238 +2606,3 @@ def test_smooth_around_connection_point_of_two_lines(
     assert equals_exact(result_1, expected_1, tolerance=0.000000001)
 
     assert equals_exact(result_2, expected_2, tolerance=0.000000001)
-
-
-@pytest.mark.parametrize(
-    (
-        "lines_gdf",
-        "points_gdf",
-        "max_distance",
-        "expected_geometries",
-    ),
-    [
-        (
-            GeoDataFrame(
-                geometry=[
-                    LineString(
-                        [
-                            [0, 0],
-                            [10, 0],
-                        ]
-                    )
-                ]
-            ),
-            GeoDataFrame(
-                geometry=[
-                    Point(5, 0),
-                ]
-            ),
-            0.1,
-            [
-                LineString([[0, 0], [5, 0]]),
-                LineString([[5, 0], [10, 0]]),
-            ],
-        ),
-        (
-            GeoDataFrame(
-                geometry=[
-                    LineString(
-                        [
-                            [0, 0],
-                            [5, 0],
-                            [10, 0],
-                        ]
-                    )
-                ]
-            ),
-            GeoDataFrame(
-                geometry=[
-                    Point(5, 0),
-                ]
-            ),
-            0.1,
-            [
-                LineString([[0, 0], [5, 0]]),
-                LineString([[5, 0], [10, 0]]),
-            ],
-        ),
-        (
-            GeoDataFrame(
-                geometry=[
-                    LineString(
-                        [
-                            [0, 0],
-                            [10, 0],
-                        ]
-                    )
-                ]
-            ),
-            GeoDataFrame(
-                geometry=[
-                    Point(2, 0),
-                    Point(7, 0),
-                ]
-            ),
-            0.1,
-            [
-                LineString([[0, 0], [2, 0]]),
-                LineString([[2, 0], [7, 0]]),
-                LineString([[7, 0], [10, 0]]),
-            ],
-        ),
-        (
-            GeoDataFrame(
-                geometry=[
-                    LineString(
-                        [
-                            [0, 0],
-                            [10, 0],
-                        ]
-                    )
-                ]
-            ),
-            GeoDataFrame(
-                geometry=[
-                    Point(5, 0.5),
-                ]
-            ),
-            1.0,
-            [
-                LineString([[0, 0], [5, 0]]),
-                LineString([[5, 0], [10, 0]]),
-            ],
-        ),
-        (
-            GeoDataFrame(
-                geometry=[
-                    LineString(
-                        [
-                            [0, 0],
-                            [10, 0],
-                        ]
-                    )
-                ]
-            ),
-            GeoDataFrame(
-                geometry=[
-                    Point(5, 2),
-                ]
-            ),
-            1.0,
-            [
-                LineString([[0, 0], [10, 0]]),
-            ],
-        ),
-        (
-            GeoDataFrame(
-                geometry=[
-                    LineString(
-                        [
-                            [0, 0],
-                            [10, 0],
-                        ]
-                    )
-                ]
-            ),
-            GeoDataFrame(
-                geometry=[
-                    Point(0, 0),
-                    Point(10, 0),
-                ]
-            ),
-            0.1,
-            [
-                LineString([[0, 0], [10, 0]]),
-            ],
-        ),
-        (
-            GeoDataFrame(
-                geometry=[
-                    LineString(
-                        [
-                            [0, 0],
-                            [5, 0],
-                            [10, 0],
-                        ]
-                    )
-                ]
-            ),
-            GeoDataFrame(
-                geometry=[
-                    Point(5, 0),
-                    Point(5, 0),
-                ]
-            ),
-            0.1,
-            [
-                LineString([[0, 0], [5, 0]]),
-                LineString([[5, 0], [10, 0]]),
-            ],
-        ),
-        (
-            GeoDataFrame(
-                geometry=[
-                    LineString(
-                        [
-                            [0, 0],
-                            [10, 0],
-                        ]
-                    ),
-                    LineString(
-                        [
-                            [0, 10],
-                            [10, 10],
-                        ]
-                    ),
-                ]
-            ),
-            GeoDataFrame(
-                geometry=[
-                    Point(5, 0),
-                    Point(2, 10),
-                ]
-            ),
-            0.1,
-            [
-                LineString([[0, 0], [5, 0]]),
-                LineString([[5, 0], [10, 0]]),
-                LineString([[0, 10], [2, 10]]),
-                LineString([[2, 10], [10, 10]]),
-            ],
-        ),
-    ],
-    ids=[
-        "single split",
-        "split at existing vertex",
-        "multiple splits",
-        "snap nearby point",
-        "point too far away",
-        "ignore endpoints",
-        "duplicate split points",
-        "multiple lines",
-    ],
-)
-def test_split_lines_by_points(
-    lines_gdf: GeoDataFrame,
-    points_gdf: GeoDataFrame,
-    max_distance: float,
-    expected_geometries: list[LineString],
-):
-    result = split_lines_by_points(
-        lines_gdf,
-        points_gdf,
-        max_distance=max_distance,
-    )
-
-    assert len(result) == len(expected_geometries)
-
-    for result_geom, expected_geom in zip(
-        result.geometry,
-        expected_geometries,
-        strict=True,
-    ):
-        assert equals_exact(
-            result_geom,
-            expected_geom,
-            tolerance=1e-9,
-        )
